refactor styles cards
