+++ client/src/components/Cards/CatalogCard/CatalogCard.tsx
@@ -1,4 +1,3 @@
-import { makeStyles, createStyles } from "@material-ui/core/styles";
 import {
   CardContent,
   Typography,
@@ -16,12 +15,14 @@ import { faCartPlus, faCalendarDay } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useHistory } from "react-router-dom";
 import { useState } from "react";
-import CustomAlert from "../Alert/CustomAlert";
-import CustomDialog from "../CustomDialog/CustomDialog";
-import FormCreator from "../FormCreator/FormCreator";
-import updateCatalog from "../../api/updateCatalog";
-import removeCatalog from "../../api/removeCatalog";
-import { useStore } from "../../pages/DrawerAppbar/DrawerAppbar";
+import CustomAlert from "../../Alert/CustomAlert";
+import CustomDialog from "../../CustomDialog/CustomDialog";
+import FormCreator from "../../FormCreator/FormCreator";
+import updateCatalog from "../../../api/updateCatalog";
+import removeCatalog from "../../../api/removeCatalog";
+import { useStore } from "../../../pages/DrawerAppbar/DrawerAppbar";
+import useStyles from "./styles";
+
 // the below type we should reuse in Catalog Explorer
 type TcatalogCard = {
   id: string;
@@ -31,73 +32,6 @@ type TcatalogCard = {
   image?: string;
   productCount: number;
 };
-const useStyles = makeStyles(() =>
-  createStyles({
-    root: {
-      height: 355,
-      borderRadius: "8px",
-      display: "flex",
-    },
-    cardContainer: {
-      width: "100%",
-      height: "100%",
-      display: "flex",
-      flexDirection: "column",
-      justifyContent: "space-between",
-      padding: "0px",
-    },
-    media: {
-      height: 300,
-      display: "flex",
-      flexDirection: "column",
-      justifyContent: "center",
-      alignItems: "center",
-      color: "#6A5DF9",
-      gap: "8px",
-      textDecoration: "none",
-      cursor: "pointer",
-    },
-    mediaContent: {
-      display: "flex",
-      flexDirection: "column",
-      justifyContent: "center",
-    },
-    headerContainer: {
-      width: "100%",
-      display: "flex",
-      justifyContent: "space-between",
-      boxSizing: "border-box",
-    },
-    footerContainer: {
-      width: "100%",
-      display: "flex",
-      flexDirection: "column",
-      justifyContent: "flex-start",
-    },
-    typography: {
-      fontWeight: 200,
-    },
-    products: {
-      color: "#6A5DF9",
-      fontSize: "14px",
-      marginBottom: "5px",
-      cursor: "pointer",
-    },
-    createdAt: {
-      display: "flex",
-      alignItems: "center",
-      color: "grey",
-      fontWeight: "lighter",
-      fontSize: "14px",
-      gap: "8px",
-    },
-    link: {
-      width: "100%",
-      textDecoration: "none",
-      color: "inherit",
-    },
-  })
-);
 
 export default function CatalogCard({
   id,
@@ -161,10 +95,9 @@ export default function CatalogCard({
     ) : null;
   return (
     <>
-      <Card className={classes.root}>
+      <Card className={classes.catalogCard}>
         <CardContent className={classes.cardContainer}>
           <CardHeader
-            className={classes.headerContainer}
             action={
               <>
                 <IconButton onClick={openOptions} aria-label="settings">
@@ -227,7 +160,7 @@ export default function CatalogCard({
               />
             )
           )}
-          <CardContent className={classes.footerContainer}>
+          <CardContent>
             <Typography
               className={classes.products}
               onClick={() => {
